Update for Login and Register andpoint
Authenticate/Login and Authenticate/Register
